requirements.txt
python-telegram-bot==9.0.0
requests==2.18.4
youtube-dl==2018.3.3
pafy==0.5.4
wit==5.1.0